+++ sdk/personalizer/azure-ai-personalizer/src/main/java/com/azure/ai/personalizer/package-info.java
@@ -3,7 +3,7 @@
 
 /**
  * Package containing classes for creating a
- * {@link com.azure.ai.personalizer.PersonalizerAdminClient} and
+ * {@link com.azure.ai.personalizer.PersonalizerAdministrationClient} and
  * {@link com.azure.ai.personalizer.PersonalizerClient} to perform operations on
  * Azure Personalizer.
  */