+++ sdk/personalizer/azure-ai-personalizer/src/test/java/com/azure/ai/personalizer/ConfigurationTests.java
@@ -30,14 +30,14 @@ public class ConfigurationTests extends PersonalizerTestBase {
             .setRewardWaitTime(newExperimentalUnitDuration)
             .setExplorationPercentage(newExplorationPercentage)
             .setLogRetentionDays(Integer.MAX_VALUE);
-        PersonalizerAdminClient client = getAdministrationClient(httpClient, serviceVersion, true);
+        PersonalizerAdministrationClient client = getAdministrationClient(httpClient, serviceVersion, true);
         testUpdateProperties(client, properties);
         testGetProperties(client, properties);
         updateAndGetPolicy(client);
         resetPolicy(client);
     }
 
-    private void testGetProperties(PersonalizerAdminClient client, PersonalizerServiceProperties properties) {
+    private void testGetProperties(PersonalizerAdministrationClient client, PersonalizerServiceProperties properties) {
         PersonalizerServiceProperties result = client.getProperties();
         assertEquals(properties.getDefaultReward(), result.getDefaultReward());
         assertTrue(Math.abs(properties.getExplorationPercentage() - result.getExplorationPercentage()) < 1e-3);
@@ -46,7 +46,7 @@ public class ConfigurationTests extends PersonalizerTestBase {
         assertEquals(properties.getRewardWaitTime(), result.getRewardWaitTime());
     }
 
-    private void testUpdateProperties(PersonalizerAdminClient client, PersonalizerServiceProperties properties) {
+    private void testUpdateProperties(PersonalizerAdministrationClient client, PersonalizerServiceProperties properties) {
         PersonalizerServiceProperties result = client.updateProperties(properties);
         assertEquals(properties.getDefaultReward(), result.getDefaultReward());
         assertTrue(Math.abs(properties.getExplorationPercentage() - result.getExplorationPercentage()) < 1e-3);
@@ -55,7 +55,7 @@ public class ConfigurationTests extends PersonalizerTestBase {
         assertEquals(properties.getRewardWaitTime(), result.getRewardWaitTime());
     }
 
-    private void updateAndGetPolicy(PersonalizerAdminClient client) {
+    private void updateAndGetPolicy(PersonalizerAdministrationClient client) {
         PersonalizerPolicy newPolicy = new PersonalizerPolicy()
             .setName("app1")
             .setArguments("--cb_explore_adf --quadratic GT --quadratic MR --quadratic GR --quadratic ME --quadratic OT --quadratic OE --quadratic OR --quadratic MS --quadratic GX --ignore A --cb_type ips --epsilon 0.2");
@@ -68,7 +68,7 @@ public class ConfigurationTests extends PersonalizerTestBase {
         assertEquals(newPolicy.getArguments(), policy.getArguments().substring(0, length));
     }
 
-    private void resetPolicy(PersonalizerAdminClient client) {
+    private void resetPolicy(PersonalizerAdministrationClient client) {
         PersonalizerPolicy policy = client.resetPolicy();
         assertEquals("--cb_explore_adf --epsilon 0.2 --power_t 0 -l 0.001 --cb_type mtr -q ::",
             policy.getArguments());